requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.0
Django==2.2.24
django-haystack==3.1.1
django-redis==4.10.0
django-tinymce==2.7.1
djangoajax==3.2
dnspython==1.16.0
eventlet==0.31.0
greenlet==0.4.15
idna==2.8
importlib-metadata==4.8.2
itsdangerous==1.1.0
jieba==0.39
kombu==4.6.4
monotonic==1.5
mysqlclient==1.4.4
Pillow==8.3.2
pipreqs==0.4.11
py3Fdfs==2.2.0
pycparser==2.21
pycryptodomex==3.9.4
PyMySQL==0.9.3
pyOpenSSL==21.0.0
python-alipay-sdk==3.0.4
pytz==2019.2
redis==3.3.8
requests==2.26.0
simpleui==3.9
six==1.12.0
sqlparse==0.3.0
typing_extensions==4.0.1
urllib3==1.26.7
vine==1.3.0
Whoosh==2.7.4
wincertstore==0.2
zipp==3.6.0
